clean up some warnings, add a FIXME for projecting while writing PointRecord, which needs to be descaled
--HG--
extra : convert_revision : svn%3A1766ff46-f334-0410-ab20-d63176f87757/trunk%401069

+++ src/lassrs.cpp
@@ -273,7 +273,7 @@ void LASSRS::ResetVLRs()
                                       "Upgrade your libgeotiff to a version greater "
                                       "than r1531 (libgeotiff 1.2.6)");
                      }
-         uint32_t length = acount;
+         uint16_t length = acount;
          record.SetRecordLength(length);
          
          // Copy the data into the data vector
